+++ test/runtests.jl
@@ -1,6 +1,6 @@
 #                            commit 9.2 8.3 8.3 6cb 6cb 8.6 6cb 6cb
 #                           machine tig ai5 ai4 tr5 tr4 aws osx os4
-@time include("kptr.jl")          #  16   1   1   0   0  20   0   0
+#@time include("kptr.jl")          #  16   1   1   0   0  20   0   0
 @time include("gpu.jl")           #   6   1   1   0   0   2   0   0
 @time include("distributions.jl") #   2   1   1   2   1   3   3   2
 @time include("dropout.jl")       #   5                   2